# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: SmartPresence API
  description: API for managing users, rooms, classes, sessions, and attendance in an educational institution
  version: 1.0.0
  contact:
    name: SmartPresence Development Team
    email: dev@smartpresence.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.smartpresence.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: Get all rooms
      description: Retrieve all rooms (Admin only)
      responses:
        '200':
          description: Rooms retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomsResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Rooms
      summary: Create room
      description: Create a new room (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{id}:
    get:
      tags:
        - Rooms
      summary: Get room by ID
      description: Retrieve a specific room by ID (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Room retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Rooms
      summary: Update room
      description: Update an existing room (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Rooms
      summary: Delete room
      description: Delete a room (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Room deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /classes:
    get:
      tags:
        - Classes
      summary: Get all classes
      description: Retrieve all classes
      responses:
        '200':
          description: Classes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassesResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Classes
      summary: Create class
      description: Create a new class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassRequest'
      responses:
        '201':
          description: Class created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /students/attendance/mark:
    post:
      tags:
        - Attendance
      summary: Mark attendance
      description: Mark attendance for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkAttendanceRequest'
      responses:
        '201':
          description: Attendance marked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecordResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
        - path
        - method
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - message
            - type
            - statusCode
          properties:
            message:
              type: string
            type:
              type: string
            statusCode:
              type: integer
            details:
              type: array
              items:
                type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        method:
          type: string
        requestId:
          type: string

    # User Schemas
...
    Room:
      type: object
      required:
        - room_id
        - name
        - wifi_ssid
        - created_at
      properties:
        room_id:
          type: integer
        name:
          type: string
        wifi_ssid:
          type: string
        bluetooth_beacon_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    RoomResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Room'
    RoomsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Room'
    CreateRoomRequest:
      type: object
      required:
        - name
        - wifi_ssid
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        wifi_ssid:
          type: string
          minLength: 1
          maxLength: 100
        bluetooth_beacon_id:
          type: string
          maxLength: 100
    UpdateRoomRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        wifi_ssid:
          type: string
          minLength: 1
          maxLength: 100
        bluetooth_beacon_id:
          type: string
          maxLength: 100

    # Class Schemas
    Class:
      type: object
      required:
        - class_id
        - name
        - course_code
        - teacher_id
        - created_at
      properties:
        class_id:
          type: integer
        name:
          type: string
        course_code:
          type: string
        description:
          type: string
          nullable: true
        teacher_id:
          type: integer
        created_at:
          type: string
          format: date-time
    ClassResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Class'
    ClassesResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Class'
    CreateClassRequest:
      type: object
      required:
        - name
        - course_code
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
        course_code:
          type: string
          minLength: 1
          maxLength: 50
        description:
          type: string

    # Session Schemas
...
    AttendanceRecord:
      type: object
      required:
        - attendance_id
        - student_id
        - session_id
        - status
        - marked_at
      properties:
        attendance_id:
          type: integer
        student_id:
          type: integer
        session_id:
          type: integer
        status:
          type: string
          enum: [present, absent, late, excused]
        marked_at:
          type: string
          format: date-time
    AttendanceRecordResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AttendanceRecord'
...
    MarkAttendanceRequest:
      type: object
      required:
        - class_id
        - session_id
        - wifi_ssid
        - device_id
      properties:
        class_id:
          type: integer
          minimum: 1
        session_id:
          type: integer
          minimum: 1
        wifi_ssid:
          type: string
          minLength: 1
          maxLength: 100
        bluetooth_beacon_id:
          type: string
          maxLength: 100
        device_id:
          type: string
          minLength: 3
          maxLength: 200
...